Updated CI workflow with Go setup
The Continuous Integration (CI) workflow has been updated to include the setup of Go. This is done using the 'actions/setup-go' action, which is configured to use the version specified in the 'go.mod' file. The addition of this step ensures that tests are run in an environment with the correct version of Go installed.