Setup goreleaser (#40)
* Setup goreleaser

* Defines separate build configurations

Configures separate build configurations in `.goreleaser.yaml`
to allow building `devctl` and `mk_funcs` with distinct settings.

This change enhances the build process, providing flexibility for
building different components of the project.

* Removes binary name from build config

Removes the explicit binary name declaration in the build configuration.

This change allows the build process to automatically determine the output binary name based on the project structure and configuration. It simplifies the configuration and reduces the risk of naming conflicts.

* Enables CGO for mk_funcs build

Sets the CGO_ENABLED environment variable to 1 for the mk_funcs build target.

This is required to properly build and link certain C code dependencies
used by the mk_funcs build.

* Adds ARM64 architecture support

Adds ARM64 architecture support to the build process for both standard binaries and C-shared libraries.
This allows the application to run on ARM64-based systems.

* Excludes arm64 architecture from build targets

Excludes the arm64 architecture from the list of build targets.

This change reduces build times and artifact size by limiting
the number of architectures being built for the darwin platform.

* Removes redundant OS/ARCH definitions

Simplifies the build configuration by removing
the explicit definitions for operating systems and
architectures, allowing the toolchain to determine
these automatically.

* Just linux then dammit

* Idk anymore man