Adds tool installation (#45)
* Install cmd

* Durdling

* gomodtidy

* Adds install command

Adds the install command to load and display configured tools.

This change introduces the initial implementation for installing tools defined in the configuration.
It loads the workspace and configuration, then iterates through the tools, displaying their names.

* Simplifies config loading and initialization

Refactors config loading and initialization to streamline the process.

- Removes the need to explicitly handle "not found" errors during
  config loading.
- Simplifies the initialization process by directly writing the
  default configuration.

* Adds config verification

Adds a basic config verification command with e2e test.

Also adds a default logger to viper and allows reading configs
from the current directory.

* Enables tool installation from URL

Adds functionality to install tools directly from a provided URL,
either from a direct binary download or a tarball.

The install command now accepts a tool name and URL as arguments.
It downloads the tool, extracts it if necessary, and places it
in the designated binary directory.